pyproject.toml
[project]
name = "pyai"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Samuel Boyes", email = "sjboyes96@gmail.com" }
]
dependencies = [
    "openai>=1.13.3",
    "python-dotenv>=1.0.1",
    "transformers>=4.38.2",
    "torch>=2.2.1",
    "accelerate>=0.28.0",
    "marqo>=3.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"pyai" = "pyai:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pyai"]
